correct file names in inject data
Former-commit-id: ad6035614c44bc398a01f0a4b9661067ec249293

+++ modules/Platform/edm_qm/WindowsFunctionsForLinux.h
@@ -2,6 +2,6 @@
    Functions in Windows that does not exist in Linux is available here.
 *=================================================================================================*/
 
-unsigned int GetTickCount();
+unsigned long GetTickCount();
 
 int _mkdir(char * _Path);
